simplifyZoteroNotes
